requirements.txt
Flask==3.0.3
flask_sqlalchemy==3.1.1
nltk==3.9.1
pandas==2.2.3
scikit_learn==1.5.2
scikit_surprise==1.1.4
